move blocks & home.js into ui directory
